POC figure container to handle the copyright and caption conditional rendering logic
