+++ src/app/components/Figure/Caption/index.test.jsx
@@ -12,4 +12,3 @@ describe('Caption', () => {
     </Caption>,
   );
 });
-
